+++ clients/admin-ui/src/features/privacy-requests/privacy-requests.slice.ts
@@ -18,7 +18,7 @@ import {
   ConfigStorageDetailsRequest,
   ConfigStorageSecretsDetailsRequest,
   DenyPrivacyRequest,
-  GetUpdloadedManualWebhookDataRequest,
+  GetUploadedManualWebhookDataRequest,
   MessagingConfigResponse,
   PatchUploadManualWebhookDataRequest,
   PrivacyRequestEntity,
@@ -308,12 +308,20 @@ export const privacyRequestApi = baseApi.injectEndpoints({
         return cloneResponse;
       },
     }),
-    getUploadedManualWebhookData: build.query<
+    getUploadedManualAccessWebhookData: build.query<
       any,
-      GetUpdloadedManualWebhookDataRequest
+      GetUploadedManualWebhookDataRequest
     >({
       query: (params) => ({
         url: `privacy-request/${params.privacy_request_id}/access_manual_webhook/${params.connection_key}`,
+      }),
+    }),
+    getUploadedManualErasureWebhookData: build.query<
+      any,
+      GetUploadedManualWebhookDataRequest
+    >({
+      query: (params) => ({
+        url: `privacy-request/${params.privacy_request_id}/erasure_manual_webhook/${params.connection_key}`,
       }),
     }),
     resumePrivacyRequestFromRequiresInput: build.mutation<any, string>({
@@ -438,12 +446,22 @@ export const privacyRequestApi = baseApi.injectEndpoints({
         body: params,
       }),
     }),
-    uploadManualWebhookData: build.mutation<
+    uploadManualAccessWebhookData: build.mutation<
       any,
       PatchUploadManualWebhookDataRequest
     >({
       query: (params) => ({
         url: `privacy-request/${params.privacy_request_id}/access_manual_webhook/${params.connection_key}`,
+        method: "PATCH",
+        body: params.body,
+      }),
+    }),
+    uploadManualErasureWebhookData: build.mutation<
+      any,
+      PatchUploadManualWebhookDataRequest
+    >({
+      query: (params) => ({
+        url: `privacy-request/${params.privacy_request_id}/erasure_manual_webhook/${params.connection_key}`,
         method: "PATCH",
         body: params.body,
       }),
@@ -457,11 +475,11 @@ export const {
   useDenyRequestMutation,
   useGetAllPrivacyRequestsQuery,
   useGetNotificationQuery,
-  useGetUploadedManualWebhookDataQuery,
   useResumePrivacyRequestFromRequiresInputMutation,
   useRetryMutation,
   useSaveNotificationMutation,
-  useUploadManualWebhookDataMutation,
+  useUploadManualAccessWebhookDataMutation,
+  useUploadManualErasureWebhookDataMutation,
   useGetStorageDetailsQuery,
   useCreateStorageMutation,
   useCreateStorageSecretsMutation,